some issues were fixed .
issue 1 - No composition found : for some train numbers the servers resonse is no composition found. so the view as updated to hide composition related info and
instead show a message that compositoin is not found.

issue 2 - Listing the wagons - instead of using HTMl  list, a function was created which takes array of the wagons and returns makes a string of the wagons concatenated.